Handle 'nil' without warnings.
The code was generating this warning:

    Use of uninitialized value in string eq
    at lib/Locale/TextDomain/OO/Extract/Xslate.pm line 138.

While fixing that with a defined check, I noticed the conditional there
needed the parentheses tweaked. It was originally written as

    if ($arity eq 'call' && ($value eq '(') || $value eq '(call)' )

or

    if ($x && $y || $z)

which because of operator precedence is really

    if ( ($x && $y) || $z)

and I think the intent was to do actually

    if ( $x && ($y || $z) )

so I added the parentheses to actually do that, which was necessary for
the defined() check I added to work.